--- src/common/components/InputField/ErrorMessage/ErrorMessage.tsx
@@ -1,9 +1,0 @@
-import styles from "./ErrorMessage.module.scss";
-
-interface ErrorMessageProps {
-  errorMessage?: string;
-}
-
-export default function ErrorMessage({ errorMessage }: ErrorMessageProps) {
-  return <div className={styles.errorMessage}>{errorMessage}</div>;
-}
