+++ src/common/components/InputField/index.ts
@@ -1,4 +1,5 @@
-export { default as SuffixIcon } from "./SuffixIcon/SuffixIcon";
-export { default as PrefixIcon } from "./PrefixIcon/PrefixIcon";
-export { default as ErrorMessage } from "./ErrorMessage/ErrorMessage";
-export { default as Label } from "./Label/Label";
+export { default as SuffixIcon } from "./parts/SuffixIcon";
+export { default as PrefixIcon } from "./parts/PrefixIcon";
+export { default as ErrorMessage } from "./parts/ErrorMessage";
+export { default as Label } from "./parts/Label";
+export { default as Input } from "./parts/Input";
