+++ src/common/components/InputField/parts/ErrorMessage.tsx
@@ -1,0 +1,12 @@
+import classNames from "classnames/bind";
+import styles from "./ErrorMessage.module.scss";
+
+const cn = classNames.bind(styles);
+
+interface ErrorMessageProps {
+  message: string;
+}
+
+export default function ErrorMessage({ message }: ErrorMessageProps) {
+  return <div className={cn("errorMessage")}>{message}</div>;
+}
