+++ src/common/components/InputField/parts/Input.module.scss
@@ -2,6 +2,7 @@
   border: none;
   outline: none;
   @include sopoqa-16-400;
+  width: 90%;
 
   &::placeholder {
     color: $color-gray-40;
