+++ src/common/components/InputField/parts/Label.tsx
@@ -1,4 +1,7 @@
+import classNames from "classnames/bind";
 import styles from "./Label.module.scss";
+
+const cn = classNames.bind(styles);
 
 interface LabelProps {
   name: string;
@@ -7,7 +10,7 @@ interface LabelProps {
 
 export default function Label({ name, label }: LabelProps) {
   return (
-    <label htmlFor={name} className={styles.label}>
+    <label htmlFor={name} className={cn("label")}>
       {label}
     </label>
   );
