+++ src/common/components/InputField/parts/SuffixIcon.module.scss
