+++ src/common/components/index.ts
@@ -1,2 +1,2 @@
 export { default as Button } from "./Button/Button";
-export { default as Input } from "./InputField/Input/Input";
+export { default as InputField } from "./InputField/InputField";
